+++ app/javascript/vue/tasks/collecting_events/filter/app.vue
@@ -9,6 +9,7 @@
       :selected-ids="selectedIds"
       :url-request="urlRequest"
       :list="list"
+      :csv-options="csvOptions"
       v-model:append="append"
       @filter="makeFilterRequest({ ...parameters, extend, page: 1 })"
       @per="makeFilterRequest({ ...parameters, extend, page: 1 })"
@@ -94,7 +95,7 @@ import FilterComponent from './components/Filter.vue'
 import MapComponent from './components/Map.vue'
 import FilterLayout from '@/components/layout/Filter/FilterLayout.vue'
 import VSpinner from '@/components/ui/VSpinner.vue'
-import useFilter from '@/shared/Filter/composition/useFilter.js'
+import { useFilter, useCSVOptions } from '@/shared/Filter/composition'
 import RadialCollectingEvent from '@/components/radials/ce/radial.vue'
 import FilterList from '@/components/Filter/Table/TableResults.vue'
 import TableLayoutSelector from '@/components/Filter/Table/TableLayoutSelector.vue'
@@ -169,6 +170,7 @@ const {
   initParameters: { extend }
 })
 
+const csvOptions = useCSVOptions({ layout: currentLayout, list })
 const isMouseDown = ref(false)
 const rowHover = ref()
 const georeferences = computed(() =>