+++ CHANGELOG.md
@@ -7,16 +7,30 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ## [unreleased]
 
-\-
+### Added
+
+- Offset parameter to copy_table_to_clipboard helper [#4265]
+- Collecting event metadata task (count of use, plot of time), via filter
+- UI flag that a DwcOccurrence re-index is pending. [#4267]
 
 ### Fixed
 
+- Min/max use of Collecting Event filter facet [#4252]
 - Unify Otu failing because of CachedMap references
 - Determiner is not carrying over in locked, containerized COs in CSD [#4250]
-- CO Filter - blank fields dispersed among sor [#4255]
+- Filter CO: blank fields dispersed among sort [#4255]
+- Filter CO: otu name is empty in taxon determinations rows [#4257]
+- Filter CE: some data attributes are missing from results [#4258]
+- Filters: Download CSV doesn't use selected layout to generate files [#4264]
 
+[#4252]: https://github.com/SpeciesFileGroup/taxonworks/issues/4252
 [#4250]: https://github.com/SpeciesFileGroup/taxonworks/issues/4250
 [#4255]: https://github.com/SpeciesFileGroup/taxonworks/issues/4255
+[#4257]: https://github.com/SpeciesFileGroup/taxonworks/issues/4257
+[#4258]: https://github.com/SpeciesFileGroup/taxonworks/issues/4258
+[#4264]: https://github.com/SpeciesFileGroup/taxonworks/issues/4264
+[#4265]: https://github.com/SpeciesFileGroup/taxonworks/issues/4265
+[#4267]: https://github.com/SpeciesFileGroup/taxonworks/issues/4267
 
 ## [0.49.0] - 2025-03-18